requirement.txt
asn1crypto==0.23.0
bcrypt==3.1.4
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
colorama==0.3.9
cryptography==2.1.4
Django==2.0.2
django-filter==1.1.0
djangorestframework==3.7.7
Flask==0.12.2
Flask-Cors==3.0.3
httpie==0.9.9
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
mysqlclient==1.3.12
paramiko==2.4.0
Pillow==5.0.0
pyasn1==0.4.2
pycparser==2.18
Pygments==2.2.0
PyNaCl==1.2.0
pytesseract==0.1.7
pytz==2018.3
requests==2.18.4
six==1.11.0
tornado==4.5.2
urllib3==1.22
Werkzeug==0.14.1
xlrd==1.1.0
